Fix slicing bug when actionChoice outputs > 65535 & update in-game costs

+++ contracts/Players/PlayersImplProcessActions.sol
@@ -251,7 +251,7 @@ contract PlayersImplProcessActions is PlayersImplBase, PlayersBase {
       emit DailyReward(
         _from,
         _playerId,
-        uint16(pendingQueuedActionState.dailyRewardItemTokenIds[0]),
+        pendingQueuedActionState.dailyRewardItemTokenIds[0],
         pendingQueuedActionState.dailyRewardAmounts[0]
       );
 
@@ -259,7 +259,7 @@ contract PlayersImplProcessActions is PlayersImplBase, PlayersBase {
         emit WeeklyReward(
           _from,
           _playerId,
-          uint16(pendingQueuedActionState.dailyRewardItemTokenIds[1]),
+          pendingQueuedActionState.dailyRewardItemTokenIds[1],
           pendingQueuedActionState.dailyRewardAmounts[1]
         );
       }
